[remove mj-page] rename mj-single.js to main.js
* update tests
* update examples in bin
* update README
