+++ README.md
@@ -1,10 +1,8 @@
 # mathjax-node [![Build Status](https://travis-ci.org/mathjax/MathJax-node.svg?branch=develop)](https://travis-ci.org/mathjax/MathJax-node)
 
-This repository contains files that provide APIs to call [MathJax](https://github.com/mathjax/mathjax) from
-node.js programs.  There is an API for converting individual math
-expressions (in any of MathJax's input formats) into SVG images or MathML
-code, and there is an API for converting HTML snippets containing any of
-MathJax input formats into HTML snippets containing SVG or MathML.
+This repository contains a library that provides an API to call [MathJax](https://github.com/mathjax/mathjax) from
+NodeJS programs.  The API converts individual math
+expressions (in any of MathJax's input formats) into HTML (with CSS), SVG or MathML code.
 
 See the comments in the individual files for more details.
 
@@ -18,18 +16,18 @@ Use
 
 to install mathjax-node and its dependencies.
 
-These API's no longer produce PNG images but can easily be integrated into workflows using svg-to-png converters (e.g., svg2png, librsvg, batik).
+**Note:**
+
+* mathjax-node no longer includes an API for processing HTML-fragments but can easily be integrated into workflows for processing larger documents.
+* the API no longer produces PNG images but can easily be integrated into workflows using svg-to-png converters (e.g., svg2png, librsvg, batik).
 
 # Getting started
 
-mathjax-node provides two libraries, `./lib/mj-single.js` and `./lib/mj-page.js`. Below are two  very minimal examples -- be sure to check out the examples in `./bin/` for more advanced configurations.
-
-* `./lib/mj-single.js` is optimized for processing single equations.
-
+mathjax-node provides a library, `./lib/main.js`. Below is a very minimal example - be sure to check out the examples in `./bin/` for more advanced examples.
 
 ```javascript
 // a simple TeX-input example
-var mjAPI = require("mathjax-node/lib/mj-single.js");
+var mjAPI = require("mathjax-node");
 mjAPI.config({
   MathJax: {
     // traditional MathJax configuration
@@ -45,30 +43,21 @@ mjAPI.typeset({
   mml:true, //  svg:true,
 }, function (data) {
   if (!data.errors) {console.log(data.mml)}
+  // will produce:
+  // <math xmlns="http://www.w3.org/1998/Math/MathML" display="block">
+  //   <mi>E</mi>
+  //   <mo>=</mo>
+  //   <mi>m</mi>
+  //   <msup>
+  //     <mi>c</mi>
+  //     <mn>2</mn>
+  //   </msup>
+  // </math>
 });
 ```
 
+## Examples
 
-* `./lib/mj-page.js` is optimized for handling full HTML pages.
+The examples in `./bin/` provide a starting point for more advanced integrations.
 
-
-```javascript
-var mjAPI = require("mathjax-node/lib/mj-page.js");
-var jsdom = require("jsdom").jsdom;
-
-var document = jsdom("<!DOCTYPE html><html lang='en'><head><title>Test</title></head><body><h1>Let's test mj-page</h1> <p> \\[f: X \\to Y\\], where \\( X = 2^{\mathbb{N}}\\) </p></body></html>");
-
-mjAPI.start();
-
-mjAPI.typeset({
-  html: document.body.innerHTML,
-  renderer: "NativeMML",
-  inputs: ["TeX"],
-  xmlns: "mml"
-}, function(result) {
-  "use strict";
-  document.body.innerHTML = result.html;
-  var HTML = "<!DOCTYPE html>\n" + document.documentElement.outerHTML.replace(/^(\n|\s)*/, "");
-  console.log(HTML);
-});
-```
+Be sure to also check out other [projects on NPM that depend on mathjax-node ](https://www.npmjs.com/browse/depended/mathjax-node).
