+++ bin/mml2svg
@@ -23,7 +23,7 @@
  *  limitations under the License.
  */
 
-var mjAPI = require("../lib/mj-single.js");
+var mjAPI = require("../lib/main.js");
 
 var argv = require("yargs")
   .demand(1).strict()
