+++ bin/mml2svg-html5
@@ -25,7 +25,7 @@
  *  limitations under the License.
  */
 
-var mjAPI = require("../lib/mj-single.js");
+var mjAPI = require("../lib/main.js");
 var fs = require('fs');
 var jsdom = require('jsdom').jsdom;
