+++ bin/tex2svg-filter
@@ -29,7 +29,7 @@
  *  limitations under the License.
  */
 
-var mjAPI = require("../lib/mj-single.js");
+var mjAPI = require("../lib/main.js");
 var readline = require('readline');
 var fs = require('fs');
 
@@ -92,7 +92,7 @@ var rl = readline.createInterface({
 var tex = [], n = 0;
 
 //
-//  Data passed to mj-single-svg to maintain state
+//  Data passed to main-svg to maintain state
 //
 var state = {};
