+++ lib/main.js
@@ -1,6 +1,6 @@
 /*********************************************************************
  *
- *  mj-single-svg.js
+ *  main.js
  *
  *  Implements an API to MathJax in Node.js so that MathJax can be
  *  used server-side to generate HTML+CSS, SVG, or MathML.
