+++ package.json
@@ -31,7 +31,7 @@
   "scripts": {
     "test": "tape test/*.js"
   },
-  "main": "./lib/mj-single.js",
+  "main": "./lib/main.js",
   "devDependencies": {
     "tape": "^4.0.3"
   }
