+++ test/issue241.js
@@ -1,5 +1,5 @@
 var tape = require('tape');
-var mjAPI = require("../lib/mj-single.js");
+var mjAPI = require("../lib/main.js");
 var jsdom = require('jsdom').jsdom;
 
 tape('SVG output: add xlink to href in <image>', function(t) {
