+++ test/svg-metadata.js
@@ -1,5 +1,5 @@
 var tape = require('tape');
-var mjAPI = require("..//lib/mj-single.js");
+var mjAPI = require("..//lib/main.js");
 var jsdom = require('jsdom').jsdom;
 
 tape('the SVG output should add dimensions and styles', function(t) {
